Rustup
* Use {:?} instead of {} if we have to use fmt::Show, as {} now
  looks for a fmt::String implementation.
* int -> isize, uint -> usize
* #![allow(unstable)] in documentation tests.
* #![feature(box_syntax)] in crates.
* HashMap::entry takes key by value.
* Hash takes a Hasher parameter.

+++ docopt_macros/examples/add.rs
@@ -5,7 +5,7 @@ extern crate "rustc-serialize" as rustc_serialize;
 extern crate docopt;
 #[plugin] extern crate docopt_macros;
 
-docopt!(Args, "Usage: add <x> <y>", arg_x: int, arg_y: int);
+docopt!(Args, "Usage: add <x> <y>", arg_x: usize, arg_y: usize);
 
 fn main() {
     let args: Args = Args::docopt().decode().unwrap_or_else(|e| e.exit());
